feat : ExpertReport.java (전문가) 사진등록 기능 추가 Fix : DB 변경에 따른 샘플 데이터 수정(위도,경도), application (dto,entity) 위도,경도 삭제

+++ src/main/java/com/green/backend/expertreport/entity/ExpertReport.java
@@ -17,7 +17,7 @@ public class ExpertReport extends BaseTime {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer treeId;
+    private Long treeId;
 
     //fk
     @ManyToOne( cascade = CascadeType.ALL , fetch = FetchType.LAZY )
